SadLoopKernel(AVX2) and SadLoopKernelHme0(AVX2/AVX512)
Fix bug where the loading of 4 data rows was done in an incorrect order, only in the case where width==4 and height%4=0.  Instead of loading (3, 1) and (2, 4), we change the behavior to load (1, 3) and (2, 4) so that mpsad with the source can be done correctly.  This order matches what is done in other places inside of SadLoopKernel, e.g. the case where width==8 and height%4=0.