EbEntropyCoding: avoid infinite loop when counting number of bits (#314)
This happens when using the EbH265EncStreamHeader API, after initializing the encoder with codeVpsSpsPps and before receiving the encoded packet.

In this case, some fields such as maxDpbSize, encoderBitDepth in SequenceControlSet that used by PacketizationKernel thread was not copied from main thread. Since maxDbpSize is 0, when creating VPS, WriteUvlc does infinite loop. 
If the bits value is 0, it cannot exit the while loop with the shift operation.

Signed-off-by: Yeongjin Jeong <yeongjin.jeong@navercorp.com>