Modified mcts test
Modified the mcts test to adjust the maximum column
and row for the tile layout to be based on the frame
size (the number of luma pixels determines max rows and
max cols).

Also moved the actual determination of the max
column and row to where the video has already been
selected for the test.

Added -tile_slice-mode and -umv parameters for the
test.

Added decoding test for mcts to verify resulting bitstream
can be read.  This is similar to existing decode tests.

MCTS tests need to be run using updated HM decoder which
has a positive return code for when the decoding values.

Signed-off-by: Mark Feldman <mark.feldman@intel.com>