Worked around the incorrect tranform issue with AVX unsupported CPU.
The issue can be reproduced in legacy CPUs without AVX intrinsic, or
by hard coding ASM_TYPES with 1. It's caused by some incorrect
implementation of Transform8x8_SSE4_1_INTRIN() for 10-bit YUV source.
(Need to check whether the issue also happens or not with 8-bit YUV
source.)

So temporarily worked around this issue, like what the function
transformFunctionTableEncode1() is invoked since v1.2.0. And did the
same change for N2 and N4 transform coefficient shapes in addition to
the default one, just in case.

Signed-off-by: Austin Hu <austin.hu@intel.com>