Fixed the encoded bitstreams mismatch issue with the "-nb" option.
With the input buffer pre-loading option enabled, the YUV frames will
be cached into the buffer list (sequenceBuffer). But the function
ReadInputFrames() of SvtHevcEncApp would repeatly go back to the first
element of the buffer list, treated as ring buffer, so that the following
frame N+1 to the end of input video stream couldn't be encoded.

So fixed the issue by resetting bufferedInput, to continue reading the
not pre-loaded frames.

Signed-off-by: Austin Hu <austin.hu@intel.com>