Fixed the Y4M header analysis issue in SvtHevcEncApp. (#380)
The issue was caused by the definition of EB_STRNCPY invoking strncpy_ss with sizeof(dst). 
sizeof() an array can return the memory size of the array, while sizeof a pointer pointing to an array would return the size of the pointer (8 in 64bit machine, and 4 in 32bit one). 
So that dmax of strncpy_ss was always 8 or 4, and would cause the byte assignment issue once the size of source was greater or equal than dmax.

Signed-off-by: Austin Hu <austin.hu@intel.com>