Fix another random hang issue with thousands (> 5000) of frames input
Encoding of a picture A's dependents may complete later, so that the
entry (where picture A locates) of referencePictureQueue has little
chance to be cleaned up which depends on refCount, so that adding new
entrie into referencePictureQueue couldn't go ahead as expected.

Also refined the object release step in EncDec, where PPCS is still
used.

Signed-off-by: Austin Hu <austin.hu@intel.com>