MaxNumReorderPics change
Changed the algorithm used to set the MaxNumReorderPics
written to the bitstream (both in SPS and VPS structures)

NumReorderPics indicates the maximum number of pictures that can precede
any picture in decoding order and follow it in output order.  The value
needs to be no greater than the DPB size

Added DetermineNumReorderPics to calculate the value based on the
prediction structure in the configuration.  For Low Delay we set
NumReorderPics to 0 and for Random Acces we set NumReorderPics to
a value according to the number of hierarchical levels

Signed-off-by: Mark Feldman <mark.feldman@intel.com>