Address review comments of mem leak.
mem leak is not handled in the calling function itself.
But some source code error is fixed, so that any throw will cause EbInitEncoder return EB_ErrorInsufficientResources
to the application.
Application should terminate encoder in this case.
This makes it easier to handle many possible mem leaks because of throw.

Signed-off-by: Jun Tian <jun.tian@intel.com>