Replaced memcpy with referencing input frame buffers.
By replacing memcpy in EbH265EncSendPicture, with direct assigning the
pointers of Input FIFO object with the ones of the EB_H265_ENC_INPUT
object YUV buffers, which should be safe (loked) by application.

The buffers belonging to an Input FIFO object don't need to be padded,
because they're just used as source sample data in PictureAnalysis,
MotionEstimation and EncDec (sharpness enabled) kernels. But, their
width and height are "required" to be aligned with LCU size, when
doing ModeDecision for the LCUs crossing over the right or bottom
boundary of input frame. So the solution temporarily applies for 64
sample aligned resolutions of input frames, such as 8K.

ffmpeg SVT-HEVC plugin: references the input AVFrame objects, and
manages the input referenced frames with doubly-linked list.

Note: GStreamer SVT-HEVC plugin should have the similar change with
      one of ffmpeg.

Signed-off-by: Austin Hu <austin.hu@intel.com>