Update refDenSrcPicture source as reconstructed buffer.
Because after handling all the segments of a picture in EncDec kernel,
and after generating the reconstructed buffer in EncodePass() when
it's referenced, the input buffer is no longer needed and can be
returnd to application. The source data for either SAO or
refDenSrcPicture preparation (when sharpness is enabled) can be from
reconstructed buffer.

Signed-off-by: Austin Hu <austin.hu@intel.com>